refactor(): rename apps service to app service
