+++ src/lib/app/public_api.ts
@@ -1,3 +1,3 @@
 export { VantageAppModule } from './app.module';
-export { VantageAppsService, IApp, AppType, IAppConfiguration } from './apps.service';
+export { VantageAppService, IApp, AppType, IAppConfiguration } from './app.service';
 export { VantageTagsService, ITag } from './tags.service';
